# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Starfield Company Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /games:
    get:
      operationId: listGames
      tags: [games]
      summary: List all games
      responses:
        "200":
          description: List of games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
    post:
      operationId: createGame
      tags: [games]
      summary: Create a new game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  /games/{gameId}:
    get:
      operationId: getGame
      tags: [games]
      summary: Get a game by ID
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
    patch:
      operationId: updateGame
      tags: [games]
      summary: Update a game
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
    delete:
      operationId: deleteGame
      tags: [games]
      summary: Delete a game
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /games/{gameId}/celestial:
    get:
      operationId: listCelestialBodies
      tags: [celestial]
      summary: List celestial bodies for a game
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of celestial bodies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CelestialBody"
    post:
      operationId: createCelestialBody
      tags: [celestial]
      summary: Create a celestial body (Moon/Sun goal)
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CelestialBodyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CelestialBody"
  /games/{gameId}/celestial/{celestialId}:
    patch:
      operationId: updateCelestialBody
      tags: [celestial]
      summary: Update a celestial body
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
        - name: celestialId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CelestialBodyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CelestialBody"
    delete:
      operationId: deleteCelestialBody
      tags: [celestial]
      summary: Delete a celestial body
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
        - name: celestialId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /games/{gameId}/kpis:
    get:
      operationId: listKpis
      tags: [kpis]
      summary: List KPIs for a game
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of KPIs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Kpi"
    post:
      operationId: createKpi
      tags: [kpis]
      summary: Create a KPI
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KpiInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kpi"
  /games/{gameId}/kpis/{kpiId}:
    patch:
      operationId: updateKpi
      tags: [kpis]
      summary: Update a KPI
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
        - name: kpiId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KpiUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kpi"
    delete:
      operationId: deleteKpi
      tags: [kpis]
      summary: Delete a KPI
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
        - name: kpiId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /games/{gameId}/kpis/{kpiId}/entries:
    get:
      operationId: listKpiEntries
      tags: [kpis]
      summary: List daily KPI entries
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
        - name: kpiId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KpiEntry"
    post:
      operationId: createKpiEntry
      tags: [kpis]
      summary: Log a daily KPI value
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
        - name: kpiId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KpiEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KpiEntry"
  /games/{gameId}/dashboard:
    get:
      operationId: getGameDashboard
      tags: [dashboard]
      summary: Get full game dashboard summary
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameDashboard"
components:
  schemas:
...
    Game:
      type: object
      required: [id, name, mission, playerCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        mission:
          type: string
          description: The North Star — mission statement
        description:
          type: ["string", "null"]
        playerCount:
          type: integer
        notionDatabaseId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    GameInput:
      type: object
      required: [name, mission, playerCount]
      properties:
        name:
          type: string
          minLength: 1
        mission:
          type: string
          minLength: 1
        description:
          type: string
        playerCount:
          type: integer
          minimum: 1
          maximum: 20
        notionDatabaseId:
          type: string
    GameUpdate:
      type: object
      properties:
        name:
          type: string
        mission:
          type: string
        description:
          type: string
        playerCount:
          type: integer
        notionDatabaseId:
          type: string
    Player:
      type: object
      required: [id, gameId, name, role, avatarColor, createdAt]
      properties:
        id:
          type: integer
        gameId:
          type: integer
        name:
          type: string
        role:
          type: string
        avatarColor:
          type: string
        createdAt:
          type: string
          format: date-time
...
    CelestialBody:
      type: object
      required: [id, gameId, type, label, description, createdAt]
      properties:
        id:
          type: integer
        gameId:
          type: integer
        playerId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [moon, sun, star, comet]
        label:
          type: string
        description:
          type: string
        targetDate:
          type: ["string", "null"]
        progress:
          type: integer
          description: 0-100 progress percent
        createdAt:
          type: string
          format: date-time
    CelestialBodyInput:
      type: object
      required: [type, label, description]
      properties:
        playerId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [moon, sun, star, comet]
        label:
          type: string
          minLength: 1
        description:
          type: string
        targetDate:
          type: string
        progress:
          type: integer
    CelestialBodyUpdate:
      type: object
      properties:
        label:
          type: string
        description:
          type: string
        targetDate:
          type: string
        progress:
          type: integer
...
    Risk:
      type: object
      required: [id, gameId, name, severity, status, createdAt]
      properties:
        id:
          type: integer
        gameId:
          type: integer
        name:
          type: string
          description: The carnivore label
        description:
          type: ["string", "null"]
        severity:
          type: string
          enum: [lurking, stalking, charging, tamed]
        likelihood:
          type: integer
          description: 1-5 likelihood stars
        impact:
          type: integer
          description: 1-5 impact stars
        status:
          type: string
          enum: [active, mitigated, eliminated]
        mitigationPlan:
          type: ["string", "null"]
        positionX:
          type: ["number", "null"]
        positionY:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
...
    Kpi:
      type: object
      required: [id, gameId, name, unit, target, current, createdAt]
      properties:
        id:
          type: integer
        gameId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        unit:
          type: string
        target:
          type: number
        current:
          type: number
        milestoneId:
          type: ["integer", "null"]
        linkedTaskIds:
          type: array
          items:
            type: integer
        trend:
          type: ["string", "null"]
          enum: [up, down, flat, null]
        createdAt:
          type: string
          format: date-time
    KpiInput:
      type: object
      required: [name, unit, target]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        unit:
          type: string
        target:
          type: number
        milestoneId:
          type: ["integer", "null"]
    KpiUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        unit:
          type: string
        target:
          type: number
        current:
          type: number
        milestoneId:
          type: ["integer", "null"]
    KpiEntry:
      type: object
      required: [id, kpiId, value, date, createdAt]
      properties:
        id:
          type: integer
        kpiId:
          type: integer
        value:
          type: number
        note:
          type: ["string", "null"]
        date:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
    KpiEntryInput:
      type: object
      required: [value, date]
      properties:
        value:
          type: number
        note:
          type: string
        date:
          type: string
          format: date
    GameDashboard:
      type: object
      required: [game, players, milestoneProgress, taskStats, activeRisks, kpiSummary, recentActivity]
      properties:
        game:
          $ref: "#/components/schemas/Game"
        players:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        milestoneProgress:
          type: object
          properties:
            total:
              type: integer
            harvested:
              type: integer
            blooming:
              type: integer
            growing:
              type: integer
            planted:
              type: integer
            totalStars:
              type: integer
          required: [total, harvested, blooming, growing, planted, totalStars]
        taskStats:
          type: object
          properties:
            total:
              type: integer
            done:
              type: integer
            blooming:
              type: integer
            sprouting:
              type: integer
            seed:
              type: integer
          required: [total, done, blooming, sprouting, seed]
        activeRisks:
          type: array
          items:
            $ref: "#/components/schemas/Risk"
        kpiSummary:
          type: array
          items:
            type: object
            required: [kpi, latestValue, progressPercent]
            properties:
              kpi:
                $ref: "#/components/schemas/Kpi"
              latestValue:
                type: ["number", "null"]
              progressPercent:
                type: number
        recentActivity:
          type: array
          items:
            type: object
            required: [type, title, timestamp]
            properties:
              type:
                type: string
              title:
                type: string
              timestamp:
                type: string
                format: date-time